feat: add folder commands with <command>.rs
