+++ src/telegram/commands/mod.rs
@@ -1,0 +1,11 @@
+mod add;
+mod start;
+mod help;
+mod cancel;
+mod general;
+
+pub use add::*;
+pub use start::*;
+pub use cancel::*;
+pub use help::*;
+pub use general::*;
