+++ src/telegram/commands/start.rs
@@ -1,0 +1,12 @@
+
+use teloxide::{
+    requests::Requester,
+    types::Message,
+    Bot,
+};
+
+pub async fn start(bot: Bot, msg: Message) -> crate::errors::Result<()> {
+    bot.send_message(msg.chat.id, "Welcome to Homework Helper!")
+        .await?;
+    Ok(())
+}
